+++ Network/Telegram/API/Bot/Object/Keyboard.hs
@@ -1,12 +1,10 @@
-module Network.Telegram.API.Bot.Object.Keyboard (Keyboard (..), Payload) where
+module Network.Telegram.API.Bot.Object.Keyboard (Keyboard (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), ToJSON (toJSON), object, withObject, (.:), (.=))
 import "base" Data.Function (($))
 import "base" Data.Functor ((<$>))
-import "base" Data.Int (Int, Int64)
 import "base" Text.Show (Show)
 
-import Network.Telegram.API.Bot.Endpoint (Endpoint (payload, endpoint), Payload, Capacity (Edit))
 import Network.Telegram.API.Bot.Object.Button (Button)
 
 data Keyboard = Inline [[Button]] deriving Show
@@ -18,10 +16,3 @@ instance FromJSON Keyboard where
 instance ToJSON Keyboard where
 	toJSON (Inline buttons) = object
 		["inline_keyboard" .= buttons]
-
-type instance Payload 'Edit Keyboard = (Int64, Int, Keyboard)
-
-instance Endpoint 'Edit Keyboard where
-	payload (chat_id, message_id, reply_markup) = object
-		["chat_id" .= chat_id, "message_id" .= message_id, "reply_markup" .= reply_markup]
-	endpoint _ = "editMessageReplyMarkup"